created an express server and added 3 api endpoints to get and post users data
