+++ package.json
@@ -12,6 +12,9 @@
   "license": "ISC",
   "dependencies": {
     "axios": "^1.3.4",
+    "cookie-parser": "^1.4.6",
+    "cors": "^2.8.5",
+    "express": "^4.18.2",
     "nodemon": "^2.0.21"
   }
 }
